# check performance latency