```
# palm detection inferences : 559
# hand landmark inferences  : 763
Palm detection round trip   : 3.7 ms
Hand landmark round trip    : 2.7 ms
Palm detection latency      : 3.7 ms
Hand landmark latency       : 3.1 ms
```